+++ Runtime/Scripts/Extensions/ListExtensions.cs
@@ -1,6 +1,6 @@
 /*
- * File Name: StringExtensions.cs
- * Description: This script is for extension functionality regarding strings
+ * File Name: ListExtensions.cs
+ * Description: This script is for extension functionality regarding lists
  * 
  * Author(s): Kokowolo, Will Lacey
  * Date Created: September 30, 2022
@@ -20,33 +20,6 @@ namespace Kokowolo.Utilities
         /************************************************************/
         #region Functions
         
-        public static void Swap<T>(this List<T> list, int indexA, int indexB)
-        {
-            T temp = list[indexA];
-            list[indexA] = list[indexB];
-            list[indexB] = temp;
-        }
-
-        public static string ToString<T>(this List<T> list, bool formatElements, bool withNewlineCharacter = false)
-        {
-            if (formatElements)
-            {
-                string str = "[ ";
-                for (int i = 0; i < list.Count - 1; i++)
-                {
-                    str += $"{list[i].ToString()}, ";
-                    if (withNewlineCharacter) str += "\n";
-                }
-                if (list.Count != 0) str += $"{list[list.Count - 1].ToString()} ]";
-                else str += "]";
-                return str;
-            }
-            else
-            {
-                return list.ToString();
-            }
-        }
-
         public static List<TOut> Cast<TIn, TOut>(this List<TIn> list) where TOut : TIn
         {
             return list.ConvertAll(value => (TOut) value);